requirements.txt
# Gmail API
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4

# Google Sheets API
gspread==5.12.0

# PDF Processing
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.1.0

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# Database
aiosqlite==0.19.0
sqlalchemy==2.0.23

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Data Processing
python-dateutil==2.8.2
regex==2023.10.3

# Logging & Monitoring
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Docker Support
gunicorn==21.2.0